refactor/#61: AuthenticationCodeHistoryRepository 추상화
