+++ src/main/java/com/tookscan/tookscan/security/repository/redis/AuthenticationCodeHistoryRepositoryImpl.java
@@ -1,0 +1,28 @@
+package com.tookscan.tookscan.security.repository.redis;
+
+import com.tookscan.tookscan.security.domain.redis.AuthenticationCodeHistory;
+import com.tookscan.tookscan.security.repository.AuthenticationCodeHistoryRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+@Repository
+@RequiredArgsConstructor
+public class AuthenticationCodeHistoryRepositoryImpl implements AuthenticationCodeHistoryRepository {
+
+    AuthenticationCodeHistoryRedisRepository authenticationCodeHistoryRedisRepository;
+
+    @Override
+    public AuthenticationCodeHistory save(AuthenticationCodeHistory authenticationCodeHistory) {
+        return authenticationCodeHistoryRedisRepository.save(authenticationCodeHistory);
+    }
+
+    @Override
+    public void deleteById(String value) {
+        authenticationCodeHistoryRedisRepository.deleteById(value);
+    }
+
+    @Override
+    public AuthenticationCodeHistory findByIdOrElseNull(String value) {
+        return authenticationCodeHistoryRedisRepository.findById(value).orElse(null);
+    }
+}
